dcc editor: moved code responsible for state management into dccwidget/state.go
